feat: add commit_format option
This adds the new option commit_format. It is either a string, in which case it resolves to one of the pre-defined
patterns in SemanticReleaseHelper.format_patterns, or it can be passed a Regexp object, in which case that becomes
the pattern through which the commit messages will be processed.